azure-neuromodels-project/src/functions/requirements.txt
azure-functions
requests
openai
azure-cosmos
aiohttp>=3.8.0
azure-ai-textanalytics
azure-kusto-data
azure-kusto-ingest
